Switched to u32 to be more semantically correct.

+++ spirv_cross/src/wrapper.hpp
@@ -128,8 +128,8 @@ extern "C"
     typedef struct ScType
     {
         spirv_cross::SPIRType::BaseType type;
-        size_t vecsize;
-        size_t columns;
+        uint32_t vecsize;
+        uint32_t columns;
         uint32_t *member_types;
         size_t member_types_size;
         uint32_t *array;
